# asm/amd64: optimizes maybeNOPPadding if conds