```
goos: linux
goarch: amd64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
cpu: AMD Ryzen 9 3950X 16-Core Processor            
                                    │   old.txt   │              new.txt              │
                                    │   sec/op    │   sec/op     vs base              │
Compilation/without_extern_cache-32   3.895m ± 1%   3.806m ± 3%  -2.29% (p=0.011 n=7)
Compilation_sqlite3/compiler-32       410.6m ± 3%   398.0m ± 2%  -3.06% (p=0.007 n=7)
geomean                               39.99m        38.92m       -2.68%

                                    │   old.txt    │              new.txt               │
                                    │     B/op     │     B/op      vs base              │
Compilation/without_extern_cache-32   664.4Ki ± 0%   664.4Ki ± 0%       ~ (p=0.710 n=7)
Compilation_sqlite3/compiler-32       49.40Mi ± 0%   49.40Mi ± 0%       ~ (p=0.620 n=7)
geomean                               5.662Mi        5.662Mi       +0.00%

                                    │   old.txt   │              new.txt              │
                                    │  allocs/op  │  allocs/op   vs base              │
Compilation/without_extern_cache-32    852.0 ± 0%    852.0 ± 0%       ~ (p=0.559 n=7)
Compilation_sqlite3/compiler-32       28.70k ± 0%   28.70k ± 0%       ~ (p=0.269 n=7)
geomean                               4.945k        4.945k       -0.00%

```